+++ data.txt
@@ -1,0 +1,6 @@
+T1todo1
+T1todo2
+D0d1 /by 06/12/2001 1800
+D0d2 /by 06/12/2002 2100
+E0e1 /from 06/12/2002 2100 /to 06/12/2002 2200
+T1todo3
