--- src/main/java/Command.java
@@ -1,7 +1,0 @@
-public interface Command {
-
-    public void execute(TaskList tasks, Ui ui, Storage storage);
-    public boolean isExit();
-    public void setLoading();
-
-}
