+++ src/main/java/smolbrain/command/InvalidCommand.java
@@ -1,11 +1,17 @@
-public class ListCommand implements Command {
+package smolbrain.command;
+
+import smolbrain.Storage;
+import smolbrain.Ui;
+import smolbrain.exception.InvalidCommandException;
+import smolbrain.task.TaskList;
+
+public class InvalidCommand implements Command {
 
     private boolean loading;
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showMessage("Here are the tasks in your list: ");
-        tasks.displayTasks(ui);
+        ui.showError(new InvalidCommandException());
     }
 
     @Override
