+++ src/main/java/smolbrain/exception/InvalidCommandException.java
@@ -1,0 +1,14 @@
+package smolbrain.exception;
+
+public class InvalidCommandException extends Exception{
+
+    public InvalidCommandException() {
+        super();
+    }
+
+    @Override
+    public String toString() {
+        return "I'm sorry, but I don't know what that means :-(";
+    }
+
+}
