+++ src/main/java/smolbrain/task/Deadline.java
@@ -1,3 +1,5 @@
+package smolbrain.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
