+++ src/main/java/smolbrain/task/Task.java
@@ -1,3 +1,5 @@
+package smolbrain.task;
+
 public class Task {
 
     protected String description;
