+++ src/main/java/smolbrain/task/TaskList.java
@@ -1,3 +1,8 @@
+package smolbrain.task;
+
+import smolbrain.Storage;
+import smolbrain.Ui;
+
 import java.io.IOException;
 import java.util.ArrayList;
