+++ src/main/java/smolbrain/task/Todo.java
@@ -1,3 +1,5 @@
+package smolbrain.task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
